Move CustomIndexArray from templates.h to custom_index_array.h
