+++ pycode/memilio-simulation/memilio/simulation/abm.cpp
@@ -19,6 +19,7 @@
 */
 
 #include "templates.h"
+#include "utils/custom_index_array.h"
 #include "abm/abm.h"
 #include "pybind11/attr.h"
 #include "pybind11/cast.h"
