+++ pycode/memilio-simulation/memilio/simulation/secir.cpp
@@ -22,6 +22,7 @@
 #include "compartments/simulation.h"
 #include "compartments/compartmentalmodel.h"
 #include "epidemiology/populations.h"
+#include "utils/custom_index_array.h"
 #include "secir/secir.h"
 #include "secir/parameter_studies.h"
 #include "secir/analyze_result.h"
